Rework init
