+++ requirements.txt
@@ -1,4 +1,5 @@
 black
 flake8
-pytest
+google
 gspread
+pytest
